Change lessThan and greaterThan condition.
It is not working properly if lessThan or greaterThan is same as lastValue.
Let assume that lessThan value is 10 and lastValue is 7.
At this situation, if object_readData get value of 10, it satisfy equal condtion
and then it will save 10 as lastvalue.
So, lastValue and lessThan are same, (lastValue > lessThan) and
(lastValue < lessThan) condition always return false.
It need to change condtion from <= to <, and from >= to >.

Signed-off-by: Changkeun IM <asarabi8282@gmail.com>